old/requirements.txt
# Requirements pour le Solana Trading Bot
# Installation: pip install -r requirements.txt

async-lru
pre-commit
solders
solana
websocket-client
python-decouple
flask
pytz

# === CORE DEPENDENCIES ===
# Base Python packages
requests>=2.31.0
aiohttp>=3.8.5
asyncio-throttle>=1.0.2

# Database et cache
cachetools>=5.3.0
aiosqlite>=0.19.0

# Cryptographie et blockchain
base58>=2.1.1
cryptography>=41.0.0
pynacl>=1.5.0

# === SOLANA SPECIFIC ===
# Client Solana principal
solana>=0.30.2
solders>=0.18.1

# Utilitaires Solana
anchorpy>=0.20.1
construct>=2.10.68

# Jupiter SDK pour les swaps optimaux
jupiter-python-sdk>=0.0.2.0

# Utilitaires supplémentaires pour Solana
bip-utils>=2.9.0

# === WEB3 ET CRYPTO ===
# Pour les calculs crypto et hash
coincurve>=18.0.0
ecdsa>=0.18.0

# === DATA ET ANALYTICS ===
# Manipulation de données
pandas>=2.0.3
numpy>=1.24.3

# Sérialisation
msgpack>=1.0.5
ujson>=5.8.0

# === MONITORING ET LOGGING ===
# Logging avancé
structlog>=23.1.0
colorlog>=6.7.0

# Monitoring système
psutil>=5.9.5
memory-profiler>=0.61.0

# === NOTIFICATIONS ===
# Discord notifications
discord-webhook>=1.3.0

# Telegram notifications
python-telegram-bot>=20.4

# === HTTP ET API ===
# Client HTTP amélioré
httpx>=0.24.1
urllib3>=2.0.4

# Rate limiting
ratelimit>=2.2.1
backoff>=2.2.1

# === CONFIGURATION ===
# Variables d'environnement
python-dotenv>=1.0.0

# Validation de configuration
pydantic>=2.1.1
jsonschema>=4.18.4

# === TESTING (OPTIONNEL) ===
# Tests unitaires
pytest>=7.4.0
pytest-asyncio>=0.21.1
pytest-mock>=3.11.1
pytest-cov>=4.1.0

# === DEVELOPMENT (OPTIONNEL) ===
# Code quality
black>=23.7.0
flake8>=6.0.0
mypy>=1.5.1

# Documentation
sphinx>=7.1.2

# === WEB DASHBOARD (OPTIONNEL) ===
# Dashboard web simple
flask>=2.3.2
flask-cors>=4.0.0
gunicorn>=21.2.0

# === PERFORMANCE (OPTIONNEL) ===
# Accélération JSON
orjson>=3.9.4

# Accélération crypto
pycryptodome>=3.18.0

# === SYSTÈME D'EXPLOITATION ===
# Windows
pywin32>=306; sys_platform == "win32"

# Unix/Linux
python-daemon>=3.0.1; sys_platform != "win32"

# === VERSIONS SPÉCIFIQUES POUR COMPATIBILITÉ ===
# Éviter les conflits de versions
setuptools>=68.0.0
wheel>=0.41.0
pip>=23.2.1

# Résoudre les dépendances crypto
typing-extensions>=4.7.1
pycparser>=2.21

# === COMMENTAIRES SUR LES DÉPENDANCES ===

# CORE (OBLIGATOIRE):
# - requests/aiohttp: Appels API HTTP
# - cachetools: Cache en mémoire
# - aiosqlite: Base de données async
# - base58: Décodage adresses Solana
# - solana: Client Solana principal

# OPTIONNEL MAIS RECOMMANDÉ:
# - pandas: Analyse de données avancée
# - psutil: Monitoring système
# - structlog: Logging structuré
# - python-dotenv: Variables d'environnement

# DÉVELOPPEMENT SEULEMENT:
# - pytest: Tests unitaires
# - black: Formatage du code
# - mypy: Vérification de types

# PRODUCTION AVANCÉE:
# - gunicorn: Serveur web pour dashboard
# - orjson: JSON ultra-rapide
# - memory-profiler: Profiling mémoire